Ensure that the keep-alive is sent for ALL pending batches when the pipeline is blocked.
The current implementation of the KeepAliveHandler will only send keep alive message for the batches are currently in the EventLoop.  There maybe many batches that live in the taskPending queue waiting for it's turn in the EventLoop. Those batches waiting in the taskPending queue will not get keep alive's sent to their channel/connection since the check for the keep alive requires state that can only be set in the EventLoop.

This change moves the state that is necessary to know when to start sending the keep alive messages to a non-blocking EventLoop. There are also some documentation updates, minor logging improvements, and renaming KeepAliveHandler -> ConnectionHandler.

Fixes #310

3.x backport of #311
