+++ src/main/java/org/logstash/beats/BeatsHandler.java
@@ -11,7 +11,6 @@ import javax.net.ssl.SSLHandshakeException;
 
 public class BeatsHandler extends SimpleChannelInboundHandler<Batch> {
     private final static Logger logger = LogManager.getLogger(BeatsHandler.class);
-    public static AttributeKey<Boolean> PROCESSING_BATCH = AttributeKey.valueOf("processing-batch");
     private final IMessageListener messageListener;
 
     public BeatsHandler(IMessageListener listener) {
@@ -23,8 +22,6 @@ public class BeatsHandler extends SimpleChannelInboundHandler<Batch> {
         if (logger.isTraceEnabled()){
             logger.trace("Channel Active");
         }
-        ctx.channel().attr(BeatsHandler.PROCESSING_BATCH).set(false);
-
         super.channelActive(ctx);
         messageListener.onNewConnection(ctx);
     }
@@ -35,17 +32,16 @@ public class BeatsHandler extends SimpleChannelInboundHandler<Batch> {
         if (logger.isTraceEnabled()){
             logger.trace("Channel Inactive");
         }
-        ctx.channel().attr(BeatsHandler.PROCESSING_BATCH).set(false);
+  
         messageListener.onConnectionClose(ctx);
     }
 
 
     @Override
     public void channelRead0(ChannelHandlerContext ctx, Batch batch) throws Exception {
-        logger.debug("Received a new payload");
-
-        ctx.channel().attr(BeatsHandler.PROCESSING_BATCH).set(true);
-
+        if(logger.isDebugEnabled()) {
+            logger.debug("Received a new payload");
+        }
         try {
             for (Message message : batch) {
                 if (logger.isDebugEnabled()) {
@@ -58,9 +54,13 @@ public class BeatsHandler extends SimpleChannelInboundHandler<Batch> {
                 }
             }
         }finally{
+            //this channel is done processing this payload, instruct the connection handler to stop sending TCP keep alive
+            ctx.channel().attr(ConnectionHandler.CHANNEL_SEND_KEEP_ALIVE).get().set(false);
+            if (logger.isDebugEnabled()) {
+                logger.debug("{}: batches pending: {}", ctx.channel().id().asShortText(),ctx.channel().attr(ConnectionHandler.CHANNEL_SEND_KEEP_ALIVE).get().get());
+            }
             batch.release();
             ctx.flush();
-            ctx.channel().attr(PROCESSING_BATCH).set(false);
         }
     }
